% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figscatter.R
\name{figscatter}
\alias{figscatter}
\title{Scatter plot}
\usage{
figscatter(data, var1, var2)
}
\arguments{
\item{data}{a data.frame}

\item{var1}{numeric column}

\item{var2}{numeric column}
}
\value{
scatter plot with a linear regression
}
\description{
Scatter plot
}
\examples{
figscatter(iris, Petal.Length, Petal.Width)
}
