% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potencia.R
\name{potencia}
\alias{potencia}
\title{Power a value}
\usage{
potencia(base, exponente)
}
\arguments{
\item{base}{numeric value}

\item{exponente}{numeric value}
}
\value{
a number base to power
}
\description{
Power a value
}
\examples{
potencia(3,4)
}
